symlinks/dirs/animeget/requirements.txt
beautifulsoup4==4.3.2
colorama==0.3.3
feedparser==5.1.3
python-dateutil==2.2
requests==2.3.0
six==1.8.0
